API breaking change to RecordSet.Matching
This commit changes the API and is not back compatible with previous
versions.  The function signature for type Match has been
changed from
     func Match(rep *Report) bool
to
     func Match(rec *Record) bool
Subsequently this change impacted the implementation for Matching
and LimitMatching to work on Records instead of Reports.  This
provided a performance increase since Matching did not have to
call Parse on every line of data, and it is more idiomatic for
clients to operate on Records.  This resulted in a data subset
taking 42sec for 27M lines of data.