Added facilities for Window. A time convolution for RecordSets
Window adds a facility to slide a time window over a recordset
and then take actions on just the records in the set.  It is designed
to work on a RecordSet that has already been sorted in chronological
order, perhaps using RecordSet.SortByTime().

This commit also makes bug corrections to LimitMatching so that
it returns the right number of records when called directly by a
client.