API breaking change for creating subsets
Transitioned the package to use interface  Matching instead of
type Match.  I also altered the syntax for this functionality from
rs.Matching(...) to rs.Subset(...)

Although this breaks utilities and examples already built on the
package I think this fundamental change was requied to make the
pacakge more extensible into the future and was specifically required
in order to address issue #20

In addition to this functionality change I am driving all future
additions and deletions to the package via Test Driven Development
so this change also included increasing test coverage from 12% to
a still embarassingly low 13.6%, but this will continue to be
improved upon.