Removed Record.Parse from the package
This is an API breaking change made prior to any further adoption.
This change is because different types of ais data may not
implement the same Header names that are found in the Marinecadastre
files and thus will fail a call to the existing Parse method.

The facilities to continue to use ParseFloat, ParseInt and
ParseTime remain and are used throughout the package and utilities
built upon it.  Users can use these tools as needed until a more
robust implementation for Parse is suggested or developed.