feature-track
This commit begins work on issue #29

Work done thus far adds the variables for a new error type
ErrEmptySet, a new concrete time Beginning and a new time.Duration
named All.  The two package time related variables implement a
convenience values for common use case to the Track function in
which a user wants to create a Track from the Beginning of the
RecordSet and include All matching records in the returned result.

The new error value is a signal back to the caller that the returned
RecordSet contains no records.

Test Driven development was set up in ais_Test.go to check for the
sempahore error and this commit is passing tests ALTHOUGH the function
does not meet a minimum viable opportunity to merge.