coverage
Increase test coverage for SubsetLimit

The remaining lines of SubsetLimit are difficult to cover because
they deal with errors writing to and flushing the created
Recrodset inside the function. Fixing this is going to require an
API breaking change so that rs2 is passed in which will give the
test routine access to change the underlying writer.