coverage-ais.go-report
Slight increase to coverage by commenting out the Report type and
its associated methods.  This type is not used in the package and
complicates the use of the package for new users coming into
the community. This simplification cleans up the API and allows
us to focus on more critical issues.