bug-valid
Corrects issue #30

The Valid() function for a set of Headers was not robust against
different file sources for AIS data. This commit removes .Valid()
from the API.  A future enhancement may add
RecordSet.Validate(format FileFormat) method.  The priority for
this feature will be determined as more users provide feedback on
the tools that will make their work more robust.