bug-unique-vessels
Corrects issue #35

Main changes are contained in the RecordSet.UniqueVessels() function
where a buffer was created to keep track of the state of the Reader
before it was read until EOF by the function.  Then the pointer for
the RecordSet is pointed at this new copy upon exiting the function.
Test coverage was also added to ensure this new functionality did not
cause breaking changes.

Regression testing revealed that changes were necessary to the
Record.Data() function in order to accurately reflect the way bytes are
written in a CSV record.  This in turn caused a breakage in the
Track testing functions which was also corrected.