bug-subset
Completes work on issue #36.

ais_test.go: alters the Track test so that it consistently fails when
calling Track which then calls SubsetLimit twice in a row.

ais.go: SubsetLimit now creates a copy of the data it reads and stores
it into a buffer.  At the end of the function it points the
RecordSet reader at this new data.  This is an inefficient solution,
but a more elegant solution is larger rewrite and would likely
require a custom implementation of the csv.Reader method or
a more raw Read implementation directly on the file system pointers
handled by RecordSet where Seek is moved back to the head of the
file after reaching EOF.