feature-header
Completes issue #38

The Headers struct included an element for a JSON encoded definition
file.  This was marginally useful, and added complexity across the
package and package testing.  This commit changes several files:
	modified:   ais.go
	modified:   ais_test.go
	modified:   example_basic_test.go
	modified:   example_distance_test.go
	modified:   functionExamples_test.go
	modified:   interaction.go
	modified:   vars_test.go

Main changes are in ais.go in which the Headers.definition was removed
from all types that referenced it or used it.

Additionally, this commit brings testing to 100% for all Headers
methods.