feature-subset
SubsetLimit was defaulting to make a copy of the read data in order
to restore the read pointer to top of the copied buffer which
allows multiple uses of the RecordSet.  However, this presented a
SERIOUS performance penalty and is now off by default.

ais.go: A new function signature for SubsetLimit provides a
multiuse bool argument to turn this feature on an off.
ais_test.go: updated to test with the new signature.  Testing needs
to be improved.